new logo, navbar, and icons
