+++ frontend/package.json
@@ -24,6 +24,7 @@
     "react-dom": "^19.0.0",
     "react-select": "^5.10.1",
     "react-virtualized": "^9.22.6",
+    "tailwind-merge": "^3.3.1",
     "zustand": "^5.0.3"
   },
   "devDependencies": {
