+++ frontend/src/app/globals.css
@@ -4,16 +4,16 @@
 
 @layer utilities {
   :root {
-    --background: #f6f6f6;
+    --background: #ffffff;
     --foreground: #ffffff;
-    --border: #dcdcdc;
+    --border: #d5d7da;
 
-    --primary-light: #1992d4;
-    --primary-dark: #0b69a3;
+    --primary-light: #1570ef;
+    --primary-dark: #175cd3;
 
     --text-lighter: #ffffff;
     --text-light: #627d98;
-    --text-dark: #334e68;
+    --text-dark: #414651;
   }
 
   * {
