+++ frontend/src/icons/Feedback.tsx
@@ -1,0 +1,26 @@
+import { SVGProps } from "react";
+
+type FeedbackProps = {
+  size?: number; // Add a size prop
+} & SVGProps<SVGSVGElement>;
+
+const Feedback = ({ color, size = 16, ...props }: FeedbackProps) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={size} // Use the size prop for width
+    height={size} // Keep the height equal to the width for a square aspect ratio
+    viewBox="0 0 20 21"
+    fill="none"
+    {...props}
+  >
+    <path
+      d="M5 6.5H10M5 10H13M5 16V18.3355C5 18.8684 5 19.1348 5.10923 19.2716C5.20422 19.3906 5.34827 19.4599 5.50054 19.4597C5.67563 19.4595 5.88367 19.2931 6.29976 18.9602L8.6852 17.0518C9.1725 16.662 9.4162 16.4671 9.6875 16.3285C9.9282 16.2055 10.1844 16.1156 10.4492 16.0613C10.7477 16 11.0597 16 11.6837 16H14.2C15.8802 16 16.7202 16 17.362 15.673C17.9265 15.3854 18.3854 14.9265 18.673 14.362C19 13.7202 19 12.8802 19 11.2V5.8C19 4.11984 19 3.27976 18.673 2.63803C18.3854 2.07354 17.9265 1.6146 17.362 1.32698C16.7202 1 15.8802 1 14.2 1H5.8C4.11984 1 3.27976 1 2.63803 1.32698C2.07354 1.6146 1.6146 2.07354 1.32698 2.63803C1 3.27976 1 4.11984 1 5.8V12C1 12.93 1 13.395 1.10222 13.7765C1.37962 14.8117 2.18827 15.6204 3.22354 15.8978C3.60504 16 4.07003 16 5 16Z"
+      stroke={color ?? "#717680"}
+      strokeWidth="2"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+    />
+  </svg>
+);
+
+export default Feedback;
