+++ frontend/src/icons/Tutorial.tsx
@@ -1,0 +1,25 @@
+import { SVGProps } from "react";
+
+type TutorialProps = {
+  size?: number; // Add a size prop
+} & SVGProps<SVGSVGElement>;
+
+const Tutorial = ({ color, size = 16, ...props }: TutorialProps) => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={size} // Use the size prop for width
+    height={size} // Keep the height equal to the width for a square aspect ratio
+    viewBox="0 0 20 20"
+    fill="none"
+    {...props}
+  >
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M20 3V17C20 18.6569 18.6569 20 17 20H3C1.34315 20 0 18.6569 0 17V3C0 1.34315 1.34315 0 3 0H17C18.6569 0 20 1.34315 20 3ZM7.75 4.79C8.58799 4.1053 9.69253 3.84021 10.75 4.07L10.76 4.04C12.3063 4.37147 13.44 5.69608 13.5288 7.27504C13.6175 8.854 12.6394 10.2973 11.14 10.8C10.7623 10.9939 10.5179 11.3758 10.5 11.8V12C10.5 12.5523 10.0523 13 9.5 13C8.94771 13 8.5 12.5523 8.5 12V11.77C8.52806 10.4979 9.31633 9.36677 10.5 8.9C11.1515 8.69257 11.5803 8.07078 11.5427 7.38806C11.505 6.70534 11.0104 6.1345 10.34 6C9.86476 5.89843 9.36932 6.02413 9 6.34C8.65469 6.62576 8.45636 7.05179 8.46 7.5C8.46 8.05229 8.01229 8.5 7.46 8.5C6.90772 8.5 6.46 8.05229 6.46 7.5C6.45689 6.44781 6.93137 5.45103 7.75 4.79ZM10.54 15.08C10.5375 14.9462 10.5104 14.814 10.46 14.69C10.4144 14.5609 10.3391 14.4445 10.24 14.35C9.84996 13.9623 9.22004 13.9623 8.83 14.35C8.73646 14.4413 8.66172 14.55 8.61 14.67C8.56259 14.7945 8.53885 14.9268 8.54 15.06C8.54368 15.3248 8.64727 15.5784 8.83 15.77C9.01872 15.9572 9.2742 16.0615 9.54 16.06H9.73L9.92 16C9.97969 15.9761 10.0367 15.9459 10.09 15.91L10.24 15.79C10.4309 15.6029 10.5389 15.3472 10.54 15.08ZM17 2C17.5523 2 18 2.44772 18 3V17C18 17.5523 17.5523 18 17 18H3C2.44772 18 2 17.5523 2 17V3C2 2.44772 2.44772 2 3 2H17Z"
+      fill={color ?? "#717680"}
+    />
+  </svg>
+);
+
+export default Tutorial;
